+++ packages/utils/src/index.ts
@@ -41,3 +41,4 @@ export * from './urlToOnion';
 export * from './logs';
 export * from './logsManager';
 export * from './bigNumber';
+export * from './throttler';
